Remove unused scripts and update cli script
- Remove unused database scripts for other methods.
- Add history file and database files to gitignore.
- Update history script to dynamically save the database file depending
  on the history file.
- Add config for cli script for easier access of global variables.
- Fix for typper cli terminal dependency not working correctly.
- Updated pip requirements.
