+++ src/cli/config.py
@@ -1,0 +1,2 @@
+HISTORY_FILE = "../db/history.json"
+HISTORY_VERSION = 1
