+++ src/server/requirements.txt
@@ -2,9 +2,7 @@ black==24.8.0
 blinker==1.8.2
 click==8.1.7
 exceptiongroup==1.2.2
-exceptiongroup==1.2.2
 Flask==3.0.3
-Flask-SQLAlchemy==3.1.1
 greenlet==3.0.3
 iniconfig==2.0.0
 isodate==0.6.1
@@ -23,6 +21,7 @@ pyparsing==3.1.2
 pytest==8.3.2
 pytest-flask==1.3.0
 python-dotenv==1.0.1
+PyYAML==6.0.2
 rdflib==7.0.0
 rich==13.7.1
 shellingham==1.5.4
